+++ src/ripple/test/jtx.h
@@ -1,0 +1,54 @@
+//------------------------------------------------------------------------------
+/*
+    This file is part of rippled: https://github.com/ripple/rippled
+    Copyright (c) 2012, 2013 Ripple Labs Inc.
+
+    Permission to use, copy, modify, and/or distribute this software for any
+    purpose  with  or without fee is hereby granted, provided that the above
+    copyright notice and this permission notice appear in all copies.
+
+    THE  SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
+    WITH  REGARD  TO  THIS  SOFTWARE  INCLUDING  ALL  IMPLIED  WARRANTIES  OF
+    MERCHANTABILITY  AND  FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR
+    ANY  SPECIAL ,  DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
+    WHATSOEVER  RESULTING  FROM  LOSS  OF USE, DATA OR PROFITS, WHETHER IN AN
+    ACTION  OF  CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
+    OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
+*/
+//==============================================================================
+
+#ifndef RIPPLE_TEST_JTX_H_INCLUDED
+#define RIPPLE_TEST_JTX_H_INCLUDED
+
+// Convenience header that includes everything
+
+#include <ripple/test/jtx/Account.h>
+#include <ripple/test/jtx/amounts.h>
+#include <ripple/test/jtx/any.h>
+#include <ripple/test/jtx/balance.h>
+#include <ripple/test/jtx/Env.h>
+#include <ripple/test/jtx/fee.h>
+#include <ripple/test/jtx/flags.h>
+#include <ripple/test/jtx/JTx.h>
+#include <ripple/test/jtx/multisign.h>
+#include <ripple/test/jtx/noop.h>
+#include <ripple/test/jtx/offer.h>
+#include <ripple/test/jtx/owners.h>
+#include <ripple/test/jtx/paths.h>
+#include <ripple/test/jtx/pay.h>
+#include <ripple/test/jtx/rate.h>
+#include <ripple/test/jtx/regkey.h>
+#include <ripple/test/jtx/require.h>
+#include <ripple/test/jtx/requires.h>
+#include <ripple/test/jtx/sendmax.h>
+#include <ripple/test/jtx/seq.h>
+#include <ripple/test/jtx/sig.h>
+#include <ripple/test/jtx/tags.h>
+#include <ripple/test/jtx/ter.h>
+#include <ripple/test/jtx/ticket.h>
+#include <ripple/test/jtx/trust.h>
+#include <ripple/test/jtx/txflags.h>
+#include <ripple/test/jtx/utility.h>
+
+#endif
+
