+++ src/ripple/test/jtx/Env.h
@@ -1,0 +1,476 @@
+//------------------------------------------------------------------------------
+/*
+    This file is part of rippled: https://github.com/ripple/rippled
+    Copyright (c) 2012, 2013 Ripple Labs Inc.
+
+    Permission to use, copy, modify, and/or distribute this software for any
+    purpose  with  or without fee is hereby granted, provided that the above
+    copyright notice and this permission notice appear in all copies.
+
+    THE  SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
+    WITH  REGARD  TO  THIS  SOFTWARE  INCLUDING  ALL  IMPLIED  WARRANTIES  OF
+    MERCHANTABILITY  AND  FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR
+    ANY  SPECIAL ,  DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
+    WHATSOEVER  RESULTING  FROM  LOSS  OF USE, DATA OR PROFITS, WHETHER IN AN
+    ACTION  OF  CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
+    OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
+*/
+//==============================================================================
+
+#ifndef RIPPLE_TEST_JTX_ENV_H_INCLUDED
+#define RIPPLE_TEST_JTX_ENV_H_INCLUDED
+
+#include <ripple/test/jtx/Account.h>
+#include <ripple/test/jtx/amounts.h>
+#include <ripple/test/jtx/JTx.h>
+#include <ripple/test/jtx/require.h>
+#include <ripple/test/jtx/tags.h>
+
+#include <ripple/app/ledger/Ledger.h>
+#include <ripple/json/json_value.h>
+#include <ripple/json/to_string.h>
+#include <ripple/protocol/Indexes.h>
+#include <ripple/protocol/Issue.h>
+#include <ripple/protocol/RippleAddress.h>
+#include <ripple/protocol/STAmount.h>
+#include <ripple/protocol/STObject.h>
+#include <ripple/protocol/STTx.h>
+#include <beast/is_call_possible.h>
+#include <beast/unit_test/suite.h>
+#include <boost/logic/tribool.hpp>
+#include <beast/cxx14/type_traits.h> // <type_traits>
+#include <beast/cxx14/utility.h> // <utility>
+#include <functional>
+#include <tuple>
+#include <utility>
+#include <unordered_map>
+
+namespace ripple {
+namespace test {
+namespace jtx {
+
+/** A view to an account's account root. */
+class AccountInfo
+{
+private:
+    Account account_;
+    std::shared_ptr<Ledger> ledger_;
+    boost::optional<SLE const> root_;
+
+public:
+    AccountInfo(Account const& account,
+            std::shared_ptr<Ledger> ledger)
+        : account_(account)
+        , ledger_(std::move(ledger))
+        , root_(ledger_->fetch(
+            getAccountRootIndex(account.id())))
+    {
+    }
+
+    STAmount
+    balance (Issue const& issue) const;
+
+    std::uint32_t
+    seq() const;
+
+    std::uint32_t
+    flags() const;
+};
+
+//------------------------------------------------------------------------------
+
+namespace detail {
+
+#ifdef _MSC_VER
+
+// Workaround for C2797:
+// list initialization inside member initializer
+// list or non-static data member initializer is
+// not implemented
+//
+template <std::size_t N>
+struct noripple_helper
+{
+    std::array<Account, N> args;
+
+    template <class T, class ...Args>
+    std::array<Account, sizeof...(Args)>
+    flatten (Args&& ...args)
+    {
+        return std::array<T,
+            sizeof...(Args)> {
+                std::forward<Args>(args)...};
+    }
+
+    template <class... Args>
+    explicit
+    noripple_helper(Args const&... args_)
+        : args(flatten<Account>(args_...))
+    {
+    }
+};
+
+#else
+
+template <std::size_t N>
+struct noripple_helper
+{
+    std::array<Account, N> args;
+
+    template <class... Args>
+    explicit
+    noripple_helper(Args const&... args_)
+        : args { args_... }
+    {
+    }
+};
+
+#endif
+
+} // detail
+
+/** Designate accounts as no-ripple in Env::fund */
+template <class... Args>
+detail::noripple_helper<1 + sizeof...(Args)>
+noripple (Account const& account,
+    Args const&... args)
+{
+    return detail::noripple_helper<
+        1 + sizeof...(Args)>(
+            account, args...);
+}
+
+//------------------------------------------------------------------------------
+
+/** A transaction testing environment. */
+class Env
+{
+public:
+    beast::unit_test::suite& test;
+
+    /** The master account. */
+    Account const master;
+
+    /** The open ledger. */
+    std::shared_ptr<Ledger> ledger;
+
+public:
+    Env (beast::unit_test::suite& test_);
+
+    /** Associate AccountID with account. */
+    void
+    memoize (Account const& account);
+
+    /** Returns the Account given the AccountID. */
+    /** @{ */
+    Account const&
+    lookup (std::string const& base58ID) const;
+
+    Account const&
+    lookup (ripple::Account const& id) const;
+    /** @} */
+
+    /** Returns info on an Account. */
+    // DEPRECATED
+    /** @{ */
+    AccountInfo
+    info (Account const& account) const
+    {
+        return AccountInfo(account, ledger);
+    }
+
+    AccountInfo
+    operator[](Account const& account) const
+    {
+        return info(account);
+    }
+    /** @} */
+
+    /** Return an account root.
+        @return empty if the account does not exist.
+    */
+    std::shared_ptr<SLE const>
+    le (Account const& account) const;
+
+    /** Return a ledger entry.
+        @return empty if the ledger entry does not exist
+    */
+    // VFALCO NOTE Use Keylet here
+    std::shared_ptr<SLE const>
+    le (uint256 const& key) const;
+
+    /** Set the fee autofill setting. */
+    void auto_fee (bool value)
+    {
+        fill_fee_ = value;
+    }
+
+    /** Set the sequence number autofill setting. */
+    void auto_seq (bool value)
+    {
+        fill_seq_ = value;
+    }
+
+    /** Set the signature autofill setting.
+        @note autofill multisigning is not supported
+    */
+    void auto_sig (bool value)
+    {
+        fill_sig_ = value;
+    }
+
+    /** Create a JTx from parameters. */
+    template <class JsonValue,
+        class... FN>
+    JTx
+    jt (JsonValue&& jv, FN const&... fN)
+    {
+        JTx jt(std::forward<JsonValue>(jv));
+        invoke(jt, fN...);
+        autofill(jt);
+        return jt;
+    }
+
+    /** Create JSON from parameters.
+        This will apply funclets and autofill.
+    */
+    template <class JsonValue,
+        class... FN>
+    Json::Value
+    json (JsonValue&&jv, FN const&... fN)
+    {
+        auto tj = jt(
+            std::forward<JsonValue>(jv),
+                fN...);
+        return std::move(tj.jv);
+    }
+
+    /** Check a set of requirements.
+        
+        The requirements are formed
+        from condition functors.
+    */
+    template <class... Args>
+    void
+    require (Args const&... args) const
+    {
+        jtx::required(args...)(*this);
+    }
+
+    /** Submit an existing JTx.
+        This calls postconditions.
+    */
+    virtual
+    void
+    submit (JTx const& jt);
+
+    /** Apply funclets and submit. */
+    /** @{ */
+    template <class JsonValue, class... FN>
+    void
+    apply (JsonValue&& jv, FN const&... fN)
+    {
+        submit(jt(std::forward<
+            JsonValue>(jv), fN...));
+    }
+
+    template <class JsonValue,
+        class... FN>
+    void
+    operator()(JsonValue&& jv,
+        FN const&... fN)
+    {
+        apply(std::forward<
+            JsonValue>(jv), fN...);
+    }
+    /** @} */
+
+private:
+    void
+    fund (bool setDefaultRipple,
+        STAmount const& amount,
+            Account const& account);
+
+    // If you get an error here it means
+    // you're calling fund with no accounts
+    inline
+    void
+    fund (STAmount const&)
+    {
+    }
+
+    void
+    fund_arg (STAmount const& amount,
+        Account const& account)
+    {
+        fund (true, amount, account);
+    }
+
+    template <std::size_t N>
+    void
+    fund_arg (STAmount const& amount,
+        detail::noripple_helper<N> const& list)
+    {
+        for (auto const& account : list.args)
+            fund (false, amount, account);
+    }
+public:
+
+    /** Create a new account with some XRP.
+
+        These convenience functions are for easy set-up
+        of the environment, they bypass fee, seq, and sig
+        settings. The XRP is transferred from the master
+        account.
+
+        Preconditions:
+            The account must not already exist
+
+        Effects:
+            The asfDefaultRipple on the account is set,
+            and the sequence number is incremented, unless
+            the account is wrapped with a call to noripple.
+
+            The account's XRP balance is set to amount.
+
+            Generates a test that the balance is set.
+
+        @param amount The amount of XRP to transfer to
+                      each account.
+
+        @param args A heterogeneous list of accounts to fund
+                    or calls to noripple with lists of accounts
+                    to fund.
+    */
+    template<class Arg, class... Args>
+    void
+    fund (STAmount const& amount,
+        Arg const& arg, Args const&... args)
+    {
+        fund_arg (amount, arg);
+        fund (amount, args...);
+    }
+
+    /** Establish trust lines.
+
+        These convenience functions are for easy set-up
+        of the environment, they bypass fee, seq, and sig
+        settings.
+    */
+    /** @{ */
+    void
+    trust (STAmount const& amount,
+        Account const& account);
+
+    template<class... Accounts>
+    void
+    trust (STAmount const& amount, Account const& to0,
+        Account const& to1, Accounts const&... toN)
+    {
+        trust(amount, to0);
+        trust(amount, to1, toN...);
+    }
+    /** @} */
+
+protected:
+    void
+    autofill_sig (JTx& jt);
+
+    virtual
+    void
+    autofill (JTx& jt);
+
+    /** Create a STTx from a JTx
+        The framework requires that JSON is valid.
+        On a parse error, the JSON is logged and
+        an exception thrown.
+        Throws:
+            parse_error
+    */
+    // VFALCO NOTE This should be <STTx const>
+    std::shared_ptr<STTx>
+    st (JTx const& jt);
+
+    inline
+    void
+    invoke (STTx& stx)
+    {
+    }
+
+    template <class F>
+    inline
+    void
+    maybe_invoke (STTx& stx, F const& f,
+        std::false_type)
+    {
+    }
+
+    template <class F>
+    void
+    maybe_invoke (STTx& stx, F const& f,
+        std::true_type)
+    {
+        f(*this, stx);
+    }
+
+    // Invoke funclets on stx
+    // Note: The STTx may not be modified
+    template <class F, class... FN>
+    void
+    invoke (STTx& stx, F const& f,
+        FN const&... fN)
+    {
+        maybe_invoke(stx, f,
+            beast::is_call_possible<F,
+                void(Env&, STTx const&)>());
+        invoke(stx, fN...);
+    }
+
+    inline
+    void
+    invoke (JTx&)
+    {
+    }
+
+    template <class F>
+    inline
+    void
+    maybe_invoke (JTx& jt, F const& f,
+        std::false_type)
+    {
+    }
+
+    template <class F>
+    void
+    maybe_invoke (JTx& jt, F const& f,
+        std::true_type)
+    {
+        f(*this, jt);
+    }
+
+    // Invoke funclets on jt
+    template <class F, class... FN>
+    void
+    invoke (JTx& jt, F const& f,
+        FN const&... fN)
+    {
+        maybe_invoke(jt, f,
+            beast::is_call_possible<F,
+                void(Env&, JTx&)>());
+        invoke(jt, fN...);
+    }
+
+    // Map of account IDs to Account
+    std::unordered_map<
+        ripple::Account, Account> map_;
+
+    bool fill_fee_ = true;
+    bool fill_seq_ = true;
+    bool fill_sig_ = true;
+};
+
+} // jtx
+} // test
+} // ripple
+
+#endif
