+++ src/ripple/test/jtx/JTx.h
@@ -1,0 +1,74 @@
+//------------------------------------------------------------------------------
+/*
+    This file is part of rippled: https://github.com/ripple/rippled
+    Copyright (c) 2012, 2013 Ripple Labs Inc.
+
+    Permission to use, copy, modify, and/or distribute this software for any
+    purpose  with  or without fee is hereby granted, provided that the above
+    copyright notice and this permission notice appear in all copies.
+
+    THE  SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
+    WITH  REGARD  TO  THIS  SOFTWARE  INCLUDING  ALL  IMPLIED  WARRANTIES  OF
+    MERCHANTABILITY  AND  FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR
+    ANY  SPECIAL ,  DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
+    WHATSOEVER  RESULTING  FROM  LOSS  OF USE, DATA OR PROFITS, WHETHER IN AN
+    ACTION  OF  CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
+    OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
+*/
+//==============================================================================
+
+#ifndef RIPPLE_TEST_JTX_JTX_H_INCLUDED
+#define RIPPLE_TEST_JTX_JTX_H_INCLUDED
+
+#include <ripple/test/jtx/requires.h>
+#include <ripple/json/json_value.h>
+#include <ripple/protocol/TER.h>
+#include <boost/logic/tribool.hpp>
+#include <functional>
+
+namespace ripple {
+namespace test {
+namespace jtx {
+
+class Env;
+
+BOOST_TRIBOOL_THIRD_STATE(use_default)
+
+/** Execution context for applying a JSON transaction.
+    This augments the transaction with various settings.
+*/
+struct JTx
+{
+    Json::Value jv;
+    boost::tribool fill_fee = boost::logic::indeterminate;
+    boost::tribool fill_seq = boost::logic::indeterminate;
+    boost::tribool fill_sig = boost::logic::indeterminate;
+    std::function<void(Env&, JTx&)> signer;
+    requires_t requires;
+    TER ter = tesSUCCESS;
+
+    JTx() = default;
+
+    JTx (Json::Value&& jv_)
+        : jv(std::move(jv_))
+    {
+    }
+
+    JTx (Json::Value const& jv_)
+        : jv(jv_)
+    {
+    }
+
+    template <class Key>
+    Json::Value&
+    operator[](Key const& key)
+    {
+        return jv[key];
+    }
+};
+
+} // jtx
+} // test
+} // ripple
+
+#endif
