+++ src/ripple/test/jtx/fee.h
@@ -1,0 +1,70 @@
+//------------------------------------------------------------------------------
+/*
+    This file is part of rippled: https://github.com/ripple/rippled
+    Copyright (c) 2012, 2013 Ripple Labs Inc.
+
+    Permission to use, copy, modify, and/or distribute this software for any
+    purpose  with  or without fee is hereby granted, provided that the above
+    copyright notice and this permission notice appear in all copies.
+
+    THE  SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
+    WITH  REGARD  TO  THIS  SOFTWARE  INCLUDING  ALL  IMPLIED  WARRANTIES  OF
+    MERCHANTABILITY  AND  FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR
+    ANY  SPECIAL ,  DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
+    WHATSOEVER  RESULTING  FROM  LOSS  OF USE, DATA OR PROFITS, WHETHER IN AN
+    ACTION  OF  CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
+    OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
+*/
+//==============================================================================
+
+#ifndef RIPPLE_TEST_JTX_FEE_H_INCLUDED
+#define RIPPLE_TEST_JTX_FEE_H_INCLUDED
+
+#include <ripple/test/jtx/Env.h>
+#include <ripple/test/jtx/tags.h>
+#include <ripple/protocol/STAmount.h>
+#include <boost/logic/tribool.hpp>
+
+namespace ripple {
+namespace test {
+namespace jtx {
+
+/** Set the fee on a JTx. */
+class fee
+{
+private:
+    STAmount v_;
+    boost::tribool b_ =
+        boost::logic::indeterminate;
+
+public:
+    explicit
+    fee (autofill_t)
+        : b_(true)
+    {
+    }
+
+    explicit
+    fee (none_t)
+        : b_(false)
+    {
+    }
+
+    explicit
+    fee (STAmount const& v)
+        : v_(v)
+    {
+        if (! isXRP(v_))
+            throw std::runtime_error(
+                "fee: not XRP");
+    }
+
+    void
+    operator()(Env const&, JTx& jt) const;
+};
+
+} // jtx
+} // test
+} // ripple
+
+#endif
