+++ src/ripple/test/jtx/impl/rate.cpp
@@ -1,0 +1,45 @@
+//------------------------------------------------------------------------------
+/*
+    This file is part of rippled: https://github.com/ripple/rippled
+    Copyright (c) 2012, 2013 Ripple Labs Inc.
+
+    Permission to use, copy, modify, and/or distribute this software for any
+    purpose  with  or without fee is hereby granted, provided that the above
+    copyright notice and this permission notice appear in all copies.
+
+    THE  SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
+    WITH  REGARD  TO  THIS  SOFTWARE  INCLUDING  ALL  IMPLIED  WARRANTIES  OF
+    MERCHANTABILITY  AND  FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR
+    ANY  SPECIAL ,  DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
+    WHATSOEVER  RESULTING  FROM  LOSS  OF USE, DATA OR PROFITS, WHETHER IN AN
+    ACTION  OF  CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
+    OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
+*/
+//==============================================================================
+
+#include <BeastConfig.h>
+#include <ripple/test/jtx/rate.h>
+#include <ripple/protocol/JsonFields.h>
+#include <stdexcept>
+
+namespace ripple {
+namespace test {
+namespace jtx {
+
+Json::Value
+rate (Account const& account, double multiplier)
+{
+    if (multiplier > 4)
+        throw std::runtime_error(
+            "rate multiplier out of range");
+    Json::Value jv;
+    jv[jss::Account] = account.human();
+    jv[jss::TransferRate] = std::uint32_t(
+        1000000000 * multiplier);
+    jv[jss::TransactionType] = "AccountSet";
+    return jv;
+}
+
+} // jtx
+} // test
+} // ripple
