+++ src/ripple/test/jtx/impl/sig.cpp
@@ -1,0 +1,48 @@
+//------------------------------------------------------------------------------
+/*
+    This file is part of rippled: https://github.com/ripple/rippled
+    Copyright (c) 2012, 2013 Ripple Labs Inc.
+
+    Permission to use, copy, modify, and/or distribute this software for any
+    purpose  with  or without fee is hereby granted, provided that the above
+    copyright notice and this permission notice appear in all copies.
+
+    THE  SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
+    WITH  REGARD  TO  THIS  SOFTWARE  INCLUDING  ALL  IMPLIED  WARRANTIES  OF
+    MERCHANTABILITY  AND  FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR
+    ANY  SPECIAL ,  DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
+    WHATSOEVER  RESULTING  FROM  LOSS  OF USE, DATA OR PROFITS, WHETHER IN AN
+    ACTION  OF  CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
+    OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
+*/
+//==============================================================================
+
+#include <BeastConfig.h>
+#include <ripple/test/jtx/sig.h>
+#include <ripple/test/jtx/utility.h>
+
+namespace ripple {
+namespace test {
+namespace jtx {
+
+void
+sig::operator()(Env const&, JTx& jt) const
+{
+    if(boost::indeterminate(b_))
+    {
+        // VFALCO Inefficient pre-C++14
+        auto const account = account_;
+        jt.signer = [account](Env&, JTx& jt)
+        {
+            jtx::sign(jt.jv, account);
+        };
+    }
+    else
+    {
+        jt.fill_sig = b_;
+    }
+}
+
+} // jtx
+} // test
+} // ripple
