+++ src/ripple/test/jtx/impl/ticket.cpp
@@ -18,9 +18,36 @@
 //==============================================================================
 
 #include <BeastConfig.h>
+#include <ripple/test/jtx/ticket.h>
+#include <ripple/protocol/JsonFields.h>
 
-#include <ripple/app/tests/Common.cpp>
-#include <ripple/app/tests/Env.cpp>
-#include <ripple/app/tests/JTx.cpp>
+namespace ripple {
+namespace test {
+namespace jtx {
 
-#include <ripple/app/tests/Env_test.cpp>
+namespace ticket {
+
+namespace detail {
+
+Json::Value
+create (Account const& account,
+    boost::optional<Account> const& target,
+        boost::optional<std::uint32_t> const& expire)
+{
+    Json::Value jv;
+    jv[jss::Account] = account.human();
+    jv[jss::TransactionType] = "TicketCreate";
+    if (expire)
+        jv["Expiration"] = *expire;
+    if (target)
+        jv["Target"] = target->human();
+    return jv;
+}
+
+} // detail
+
+} // ticket
+
+} // jtx
+} // test
+} // ripple
