+++ src/ripple/unity/test.cpp
@@ -1,0 +1,43 @@
+//------------------------------------------------------------------------------
+/*
+    This file is part of rippled: https://github.com/ripple/rippled
+    Copyright (c) 2012, 2013 Ripple Labs Inc.
+
+    Permission to use, copy, modify, and/or distribute this software for any
+    purpose  with  or without fee is hereby granted, provided that the above
+    copyright notice and this permission notice appear in all copies.
+
+    THE  SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
+    WITH  REGARD  TO  THIS  SOFTWARE  INCLUDING  ALL  IMPLIED  WARRANTIES  OF
+    MERCHANTABILITY  AND  FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR
+    ANY  SPECIAL ,  DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
+    WHATSOEVER  RESULTING  FROM  LOSS  OF USE, DATA OR PROFITS, WHETHER IN AN
+    ACTION  OF  CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
+    OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
+*/
+//==============================================================================
+
+#include <BeastConfig.h>
+
+#include <ripple/test/jtx/impl/Account.cpp>
+#include <ripple/test/jtx/impl/amounts.cpp>
+#include <ripple/test/jtx/impl/balance.cpp>
+#include <ripple/test/jtx/impl/Env.cpp>
+#include <ripple/test/jtx/impl/fee.cpp>
+#include <ripple/test/jtx/impl/flags.cpp>
+#include <ripple/test/jtx/impl/multisign.cpp>
+#include <ripple/test/jtx/impl/offer.cpp>
+#include <ripple/test/jtx/impl/owners.cpp>
+#include <ripple/test/jtx/impl/paths.cpp>
+#include <ripple/test/jtx/impl/pay.cpp>
+#include <ripple/test/jtx/impl/rate.cpp>
+#include <ripple/test/jtx/impl/regkey.cpp>
+#include <ripple/test/jtx/impl/sendmax.cpp>
+#include <ripple/test/jtx/impl/seq.cpp>
+#include <ripple/test/jtx/impl/sig.cpp>
+#include <ripple/test/jtx/impl/ticket.cpp>
+#include <ripple/test/jtx/impl/trust.cpp>
+#include <ripple/test/jtx/impl/txflags.cpp>
+#include <ripple/test/jtx/impl/utility.cpp>
+
+#include <ripple/test/jtx/impl/Env_test.cpp>
